refactor ingredient
essentially split ingredient into two classes, ingredient now only containts an id, name and category, whereas drink.recipe ingredient contains an ingredient, a volume and a unit. This fixed a major bug in creating new drinks. the volume would overwrite the volume of all drinks containing the ingredient. now drink contains an array of recipeIngredients, which contain the ingredient, the volume and the unit.